Update AlyJl0ddcPs.md
Edited raw and coded raw transcript for formatting, punctuation, compared transcript to video to make it verbatim (raw omitted mainly a few conjunctions, prepositions and possessive clauses).

A note: the raw was transcribed in British English whereas my system is set to American English, this should explain any discrepancies in spelling between the raw and my edit though I did my best to retain the native transcribed language. 

The text body used for this edit is the same as the one used in my previous pull request for the .txt for the same video.